m68k/flashlib/flash_filter.c(elf_flash_plugin_load): Fix buffer overflow.
when plugin was not found in current directory, a new path was built
into name, allocated as strdup(name) and then extended with multiple
strcat().  That's plain wrong, because we write into unallocated
memory.  Compute rather first the total length needed, allocate it
with malloc(), and fill it with one sprintf.

Signed-off-by: Philippe De Muyter <phdm@macqel.be>