m68k/driver/linux/Makefile: fixed for compilation on linux-2.6
In m68k/driver/linux/Makefile, detection of current running kernel did not
work on 2.6.22, and hence detection that it was a 2.6 kernel was also wrong,
leading to try to use a non-existent Rules.make file.  Real compilation
of bdm driver kernel module did consequently not even start.

Fix that.

Signed-off-by: Philippe De Muyter <phdm@macqel.be>